refactor: deduplicate useless code, remove some comment , change double for float

+++ src/core/hittable.hpp
@@ -6,5 +6,5 @@ struct HitRecord;
 class Hittable {
 public:
     virtual ~Hittable() noexcept = default;
-    virtual bool hit(const ray& r, double ray_tmin, double ray_tmax, HitRecord& rec) const = 0;
+    virtual bool hit(const ray& r, float ray_tmin, float ray_tmax, HitRecord& rec) const = 0;
 };
