+++ src/shape/sphere.cpp
@@ -3,21 +3,21 @@
 #include <algorithm>
 #include <cmath>
 
-#include "core/ray.hpp"
 #include "core/hitrecord.hpp"
+#include "core/ray.hpp"
 #include "maths/vector3.hpp"
 
-sphere::sphere(const point3& center, double radius)
-    : center(center), radius(std::max(1e-9, radius)) {}
+sphere::sphere(const point3& center, float radius)
+    : center(center), radius(std::max(1e-9f, radius)) {}
 
-bool sphere::hit(const ray& r, double ray_tmin, double ray_tmax, HitRecord& rec) const {
+bool sphere::hit(const ray& r, float ray_tmin, float ray_tmax, HitRecord& rec) const {
     vector3 oc = r.origin() - center;
     auto a = r.direction().length_squared();
     auto half_b = dot(oc, r.direction());
     auto c = oc.length_squared() - radius * radius;
 
     auto discriminant = half_b * half_b - a * c;
-    if (discriminant < 0)
+    if (discriminant < 0.0f)
         return false;
     auto sqrtd = std::sqrt(discriminant);
